Complete classes.js for Task 6
This is the final commit for this Sprint Challenge.  Overall, it was not too difficult.  Only issue was not fully remembering the exact syntax for a few things.  I was able to get through the stretch goals for this challenge.  I'm not sure I did the Class stretch exactly as intended,  but the results are what they should be.